better fail safe

+++ src/main/java/net/warvale/uhcmeetup/UHCMeetup.java
@@ -109,7 +109,9 @@ public class UHCMeetup extends JavaPlugin {
     private void initLobby() {
 
         //fail safe
-        FileUtils.deleteDirectory(new File("lobby"));
+        if (new File("lobby").exists()) {
+            FileUtils.deleteDirectory(new File("lobby"));
+        }
 
         //load the lobby schematic, and create the lobby world
         this.lobbySchematic = new File(this.getDataFolder(), "minigamelobby.schematic");
